third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-setrangetext-with-events.html

<!DOCTYPE html>
<script src="../../../../resources/js-test.js"></script>
<script>
    description("This tests the selection of the text field after setSelectionRange is called during a tap");
    window.onload = function() {
        if (window.eventSender) {
            doSetSelectionRange('focus');
            doSetSelectionRange('mousedown');
            doSetSelectionRange('mouseup');
            doSetSelectionRange('click');   
        }
    }

    function doSetSelectionRange(event) {
        debug(event + ' test :');
        var textfield = document.getElementById('textfield');
        textfield.setSelectionRange(0, 0);
        var tx = textfield.offsetLeft + 4;
        var ty = textfield.offsetTop + 4;

        textfield.addEventListener(event, setSelectionRange);
        eventSender.gestureTapDown(tx, ty);
        eventSender.gestureShowPress(tx, ty);
        eventSender.gestureTap(tx, ty);
        if (event == 'mousedown') {
            shouldBe('textfield.selectionStart', '0');
            shouldBe('textfield.selectionEnd', '0');
        } else {
            shouldBe('textfield.selectionStart', '0');
            shouldBe('textfield.selectionEnd', '5');
        }
        textfield.setSelectionRange(0, 0);
        textfield.removeEventListener(event, setSelectionRange);
    }
 
    function setSelectionRange(e) {
        var textfield = document.getElementById('textfield');
        textfield.setSelectionRange(0, textfield.value.length);
    }
</script>
<input type="text" value="value" id="textfield"></input>
