third_party/WebKit/LayoutTests/fast/events/touch/gesture/gesture-tap-setrangetext-with-events-expected.txt

This tests the selection of the text field after setSelectionRange is called during a tap

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
focus test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
mousedown test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 0
mouseup test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
click test :
PASS textfield.selectionStart is 0
PASS textfield.selectionEnd is 5
